Compress the generated rules by combining rules where possible
Now that the declarativeNetRequest API supports requestDomain rule
conditions[1], let's make use of that to compress the generated
rulesets where possible. Rules that previously had a urlFilter just to
match a request domain can now often be combined.

Initial testing shows an approximate 2/3 reduction[2] of the rules
generated from the EasyList filter list[3] with this technique!

1 - https://crbug.com/1207326
2 - At time of writing, converting EasyList went from 28,191 rules,
    down to just 9,592.
3 - https://easylist.to/