Replace node-gyp and RE2 code with Puppeteer
Let's make use of Puppeteer to access the
chrome.declarativeNetRequest.isRegexSupported() API. That way, we
don't need to reimplement that using RE2 and there's no need to
compile any C++ during installation.

This was made possible thanks to the recent addition of
`--headless=chrome` to Puppeteer. Previously headless mode did not
support browser extensions.

Note: For some reason, the CI runner is failing on non-Linux platforms
      for these changes. I suspect this is down to the installed
      Chromium versions on those platforms, but I don't have time to
      investigate. For now, I have just disabled those runners.